+++ src/main/java/the_dark_jumper/cannontracer/gui/utils/FormatData.java
@@ -3,11 +3,11 @@ package the_dark_jumper.cannontracer.gui.utils;
 public class FormatData {
 	public FormatData(){}
 	
-	public FormatData(int width, int offset){
+	public FormatData(float width, float offset){
 		this.width = width;
 		this.offset = offset;
 	}
 	
-	public int width;
-	public int offset;
+	public float width;
+	public float offset;
 }
